refactor: normalize names and split mixed server tests
